fixed id hashing close #17
the length isn’t the issue, it’s that my regular expression that
generates the “on this page” link wasn’t the same as the markdown
parser’s. looked at the source for the markdown parser (marked) and
used it’s regex so it matches. if we want to reduce the length of these
hashes, we’ll have to modify the markdown parser and/or use a different
one