manual ordering of sidebar sections close #54
only problem: now the sort order is part of the article filename and
therefore the URL as well. had to manually change all links. if we
change the order of an article or section, anything that links to that
article or section will have to be changed too. the search index will
also have to update to reflect the new file paths.

`grunt link-checker` (nice to have this now) was all clear, so we
should be good to go.